update-cacert-pem-certificate-bundle: change to HEAD not refs/head/master
Dont hard code the branch name in the git rev-parse,
as the branch name may not be master in the forked
repository. Use HEAD instead of refs/heads/master,
as we have just checked out ${{ github.repository_owner }}/LibreELEC.tv

We are unable to use ${{ github.ref }} as this
resolves to refs/heads/main (as the actions default
branch is main - not master.) and the script is
run from the ${{ github.repository_owner }}/actions
repository.