update_changelog.sh per branch
With this the changelog shown on the index pages of test.libreelec.tv should be branch specific. As I have no access to the server, this is based on assumptions regarding the directory layout. I would be happy to hear your comments about this. If anything has to be changed, please reply or feel free to modify as necessary. I am sure, that at least some modifications would be necessary to integrate this in the current setup.